[FEATURE] Improve error messages in TYPO3.NamingConventions.ValidVariableName
In TYPO3 direct access to $_GET and / or $_POST is not allowed.
The valid variable name sniffs detects such variables, but the error
message is not very meanful. To improve the error message,
we output another message if the variable is $_GET / $_POST.

The unit tests was extended with this new usecases, too.

Change-Id: I83c297c20d50be302a184f4c663df17408de0043
Fixes: #10535
Releases: 0.0.4