[BUGFIX] Scope/AlwaysReturnSniff complains about missing return value
If a function has a doc comment with "@return void" in it,
there must not be a return statement. With this doc comment
addition, the method has no return value.
But if there is a return value _AND_ "@return void" is defined,
there will be thrown an error anyway.

Change-Id: I46d80daf440c66e87b03e1440767635493115ebb
Fixes: #10468
Releases: 0.0.4