[BUGFIX] AlwaysReturnSniff complains about empty return for @return void
If a function/method got a @return void in doc block comment
and the function/method contains a "return;" statement,
the mentioned sniff will raise an error.

The expected behaviour should be, that no error will be raised,
because a "return;" is equal "void". Now, the sniff will
raise an error if there is something like "return null;",
"return $foo;", "return true;", and so on.

Change-Id: Iac4a55564e1eac4180c36b3af69e089984fc4d03
Fixes: #37336
Releases: 0.0.4