[BUGFIX] Make Travis-CI-Build green
In our Travis-CI-Test-Env, we run PHPCS with PEAR
standard to get a higher quality.
At the moment, the build status is failing:
http://travis-ci.org/#!/typo3-ci/TYPO3-CodeSniffer

Lets make the Travis status Green!!!!

We ignore warnings, because mostly the line length will be
throw warnings because it is not possible due to long class names.

Change-Id: Ide677b09575f9dfc72d709f5765ae1a5defdce2f
Fixes: #41164
Releases: 0.0.4