Change the folder name to TYPO3
This is not the best solution because of the generic name TYPO3 and this is just an subproject. So we have to find a better name for this standard pool. The name of the folder must be reflect in the classname of each class. Other wise the test will failed.