Rename the name of this standard
The reason for renaming is just to avoid getting confuse if someone encounter https://github.com/typo3-ci/TYPO3 and expecting the CMS but just found the PHP_CodeSniffer standard there.
To rename the standard you have to rename the folder inside the standard folder of PHPCS and in order to execute the files after renaming doing this, you have to rename the class names too to reflect the path. This commit renaming all files.
ToDo: Renaming the whole repository after merging it to master.